requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.17.2
    # via -r requirements.in
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
beautifulsoup4==4.14.3
    # via -r requirements.in
black==25.11.0
    # via -r requirements.in
build==1.3.0
    # via pip-tools
cachetools==6.2.2
    # via google-auth
certifi==2025.11.12
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.1
    # via
    #   black
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   uvicorn
cryptography==46.0.3
    # via python-jose
ecdsa==0.19.1
    # via python-jose
fastapi==0.123.9
    # via -r requirements.in
google-api-core==2.28.1
    # via google-api-python-client
google-api-python-client==2.187.0
    # via -r requirements.in
google-auth==2.41.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.1
    # via google-api-python-client
google-auth-oauthlib==1.2.3
    # via -r requirements.in
googleapis-common-protos==1.72.0
    # via google-api-core
greenlet==3.3.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httplib2==0.31.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
isort==7.0.0
    # via -r requirements.in
librt==0.6.3
    # via mypy
lxml==6.0.2
    # via
    #   -r requirements.in
    #   python-docx
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
mypy==1.19.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
oauthlib==3.3.1
    # via requests-oauthlib
packaging==25.0
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via
    #   black
    #   mypy
pillow==12.0.0
    # via reportlab
pip-tools==7.5.2
    # via -r requirements.in
platformdirs==4.5.0
    # via black
pluggy==1.6.0
    # via pytest
proto-plus==1.26.1
    # via google-api-core
protobuf==6.33.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.9.11
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.12.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pyparsing==3.2.5
    # via httplib2
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==9.0.1
    # via -r requirements.in
python-docx==1.2.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.5.0
    # via -r requirements.in
pytokens==0.3.0
    # via black
pyyaml==6.0.3
    # via uvicorn
reportlab==4.4.5
    # via -r requirements.in
requests==2.32.5
    # via
    #   google-api-core
    #   requests-oauthlib
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
six==1.17.0
    # via ecdsa
soupsieve==2.8
    # via beautifulsoup4
sqlalchemy==2.0.44
    # via
    #   -r requirements.in
    #   alembic
starlette==0.50.0
    # via fastapi
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   beautifulsoup4
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   python-docx
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
uritemplate==4.2.0
    # via google-api-python-client
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.38.0
    # via -r requirements.in
watchfiles==1.1.1
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
